Refactor code and Fixed error prompt bug
