+++ src/test/java/terminus/command/content/link/DeleteLinkCommandTest.java
@@ -1,4 +1,4 @@
-package terminus.command.link;
+package terminus.command.content.link;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -12,7 +12,6 @@ import terminus.content.Link;
 import terminus.exception.InvalidArgumentException;
 import terminus.exception.InvalidCommandException;
 import terminus.module.ModuleManager;
-import terminus.module.NusModule;
 import terminus.parser.LinkCommandParser;
 import terminus.ui.Ui;
