+++ src/test/java/terminus/parser/LinkCommandParserTest.java
@@ -6,14 +6,13 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import terminus.command.DeleteCommand;
+import terminus.command.content.DeleteCommand;
 import terminus.command.ExitCommand;
 import terminus.command.HelpCommand;
-import terminus.command.ViewCommand;
-import terminus.command.zoomlink.AddLinkCommand;
+import terminus.command.content.ViewCommand;
+import terminus.command.content.link.AddLinkCommand;
 import terminus.exception.InvalidArgumentException;
 import terminus.exception.InvalidCommandException;
-import terminus.module.NusModule;
 
 public class LinkCommandParserTest {
